requirements.txt
amqp==2.6.0
billiard==3.6.3.0
celery==4.4.4
Django==2.2.4
django-cleanup==4.0.0
django-cors-headers==3.1.0
djangorestframework==3.10.2
future==0.18.2
kombu==4.6.10
Pillow==6.2.0
pytz==2019.2
redis==3.5.3
sqlparse==0.3.0
vine==1.3.0
